+++ src/functions/unitary/integer/combo/RFactorial.java
@@ -1,13 +1,10 @@
 package functions.unitary.integer.combo;
 
-import config.Settings;
 import functions.GeneralFunction;
 import functions.unitary.UnitaryFunction;
 import tools.DefaultFunctions;
-import tools.ParsingTools;
+import tools.MiscTools;
 import tools.VariableTools;
-
-import java.util.Map;
 
 /**
  * The standard recursive definition of factorial.
@@ -35,14 +32,7 @@ public class RFactorial extends Factorial {
 			throw new UnsupportedOperationException("RFactorial has no derivative.");
 	}
 
-	@Override
-	public double evaluate(Map<Character, Double> variableValues) {
-		if (!Settings.enforceIntegerOperations)
-			throw new IllegalStateException("RFactorial cannot be used if Settings.enforceIntegerOperations is not enabled.");
-		int argument = ParsingTools.toInteger(operand.evaluate(variableValues));
-		long prod = 1;
-		for (int i = 2; i <= argument; i++)
-			prod *= i;
-		return prod;
+	public long operate(int input) {
+		return MiscTools.factorial(input);
 	}
 }