+++ tests/IntegerTest.java
@@ -1,13 +1,23 @@
 import functions.GeneralFunction;
+import functions.binary.Rand;
 import functions.binary.integer.IntegerQuotient;
 import functions.binary.integer.Modulo;
+import functions.commutative.integer.GCD;
+import functions.commutative.integer.LCM;
 import functions.special.Constant;
+import functions.special.Variable;
+import functions.unitary.piecewise.Ceil;
+import functions.unitary.piecewise.Floor;
+import functions.unitary.piecewise.Round;
+import functions.unitary.trig.normal.Sin;
 import org.junit.jupiter.api.Test;
 
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
+@SuppressWarnings("AssertEqualsBetweenInconvertibleTypes")
 public class IntegerTest {
     @Test
     void basicIntegerQuotient1() {
@@ -17,7 +27,7 @@ public class IntegerTest {
 
     @Test
     void basicIntegerQuotient2() {
-        GeneralFunction test = new IntegerQuotient(new Constant(4.999998), new Constant(25.00002));
+        GeneralFunction test = new IntegerQuotient(new Constant(4.999999999999999999999999999999999999999999999998), new Constant(25.000000000000000000000000002));
         assertEquals(5, test.evaluate(Map.of()));
     }
 
@@ -29,9 +39,47 @@ public class IntegerTest {
 
     @Test
     void basicModulo2() {
-        GeneralFunction test = new Modulo(new Constant(7.000067), new Constant(Math.pow(2, 18)));
+        GeneralFunction test = new Modulo(new Constant(7.0000000000000000000000000000000000000000000000000067), new Constant(Math.pow(2, 18)));
         assertEquals(1, test.evaluate(Map.of()));
     }
+
+    @Test
+    void gcd() {
+        assertEquals(new Constant(4), new GCD(new Constant(4), new Constant(196), new Constant(80)));
+        assertEquals(4, new GCD(new Constant(4), new Constant(196), new Constant(80)).evaluate(Map.of()));
+    }
+
+
+    @Test
+    void lcm() {
+        assertEquals(new Constant(3920), new LCM(new Constant(4), new Constant(196), new Constant(80)));
+        assertEquals(3920, new LCM(new Constant(4), new Constant(196), new Constant(80)).evaluate(Map.of()));
+    }
+
+    @Test
+    void basicFloor() {
+        GeneralFunction test = new Floor(new Sin(new Variable('x')));
+        assertEquals(0, test.evaluate(Map.of('x', 1.0)));
+    }
+
+    @Test
+    void basicCeil() {
+        GeneralFunction test = new Ceil(new Sin(new Variable('x')));
+        assertEquals(1, test.evaluate(Map.of('x', 1.0)));
+    }
+
+    @Test
+    void basicRound() {
+        GeneralFunction test = new Round(new Sin(new Variable('x')));
+        assertEquals(1, test.evaluate(Map.of('x', 1.0)));
+    }
+
+    @Test
+    void basicRand() {
+        GeneralFunction test = new Rand(new Constant(0.0), new Constant(10.0));
+        assertNotEquals(test.evaluate(Map.of()), test.evaluate(Map.of()));
+    }
+
 
 
 }